R9 Installer
- Update addon installer
- Update CTS patch
- Update mount function
- Misc fixes

TODO:
- Disable installation of OTA survival script (WIP)
- Disable sqlite optimization (WIP) (code 14 SQLITE_CANTOPEN)
- Current sqlite implementation creates an error in application database for some devices
- Disable wiping of pre-installed GApps (WIP)

+++ addon/script/addon.sh
@@ -40,6 +40,20 @@ RED='\033[0;41m'
 LRED='\033[1;31m'
 NC='\033[0m'
 
+# Root check
+id=`id`; id=`echo ${id#*=}`; id=`echo ${id%%\(*}`; id=`echo ${id%% *}`
+if [ "$id" != "0" ] && [ "$id" != "root" ]; then
+  sleep 1
+  echo ""
+  echo "You are NOT running as root..."
+  echo ""
+  sleep 1
+  echo ""
+  echo "Please type 'su' first before typing 'addon.sh'..."
+  echo ""
+  exit
+fi;
+
 # Backup
 rm -rf /sdcard/addon
 mkdir /sdcard/addon
@@ -47,12 +61,35 @@ mkdir /sdcard/addon
 # Android SDK check
 android_sdk=`getprop ro.build.version.sdk`
 
+# Mount system partition
 mount -o rw,remount / 2>/dev/null;
 mount -o rw,remount /system 2>/dev/null;
 
+# Check system mount
+MOUNTED_FS="false"
+if [ -f /system/build.prop ]; then
+  MOUNTED_FS="true"
+fi;
+if [ "$MOUNTED_FS" == "false" ]; then
+  echo "Unable to mount system. Aborting...";
+  sleep 2
+  exit 1
+fi;
+
+# Set Busybox
+BB="/data/adb/magisk/busybox"
+if [ -f "$BB" ]; then
+  MAGISK="true"
+fi;
+if [ "$MAGISK" == "false" ]; then
+  echo "Unable to find magisk busybox. Aborting...";
+  sleep 2
+  exit 1
+fi;
+
 # Installation layout
 echo "=========================="
-busybox echo -e "${RED} BiTGApps Addon Installer ${NC}"
+$BB echo -e "${RED} BiTGApps Addon Installer ${NC}"
 echo "=========================="
 PS3='Please enter your choice: '
 options=("Assistant" "Wellbeing" "Reboot" "Quit")
@@ -60,105 +97,83 @@ select opt in "${options[@]}"
 do
     case $opt in
         "Assistant")
-            busybox echo -e "${CYAN}  => Removing Pre-installed Assistant Package ${NC}"
-            rm -rf /system/priv-app/Velvet 2>/dev/null;
-            rm -rf /system/product/priv-app/Velvet 2>/dev/null;
-            busybox echo -e "${CYAN}  => Downloading Prebuilt Assistant Package ${NC}"
-            busybox wget https://github.com/BiTGApps/BiTGApps/raw/master/addon/prebuilt/prebuilt_Velvet.tar.xz -O /cache/prebuilt_Velvet.tar.xz
-            busybox echo -e "${CYAN}  => Installing Assistant Package ${NC}"
+            $BB echo -e "${CYAN}  => Downloading Prebuilt Assistant Package ${NC}"
+            curl https://bitgapps.cf/addon/VelvetPrebuilt.tar.xz > /cache/VelvetPrebuilt.tar.xz
+            $BB echo -e "${CYAN}  => Installing Assistant Package ${NC}"
             if [ "$android_sdk" == "29" ]; then
-              busybox tar -xf /cache/prebuilt_Velvet.tar.xz -C /system/product/priv-app 2>/dev/null;
+              $BB tar -xf /cache/VelvetPrebuilt.tar.xz -C /system/product/priv-app 2>/dev/null;
               chmod 0755 /system/product/priv-app/Velvet 2>/dev/null;
               chmod 0644 /system/product/priv-app/Velvet/Velvet.apk 2>/dev/null;
               chcon -h u:object_r:system_file:s0 "/system/product/priv-app/Velvet/Velvet.apk" 2>/dev/null;
-              busybox echo -e "${CYAN}  => Check Installed Package ${NC}"
+              $BB echo -e "${CYAN}  => Check Installed Package ${NC}"
               if [ -f /system/product/priv-app/Velvet/Velvet.apk ]; then
-                busybox echo -e "${CYAN}  => Package installed ${NC}"
-                if [ -f /system/product/priv-app/Velvet/Velvet.apk ]; then
-                  busybox echo -e "${CYAN}  => Backup Assistant Package ${NC}"
-                  cd /system/product/priv-app
-                  busybox tar -czf /sdcard/addon/prebuilt_Velvet.tar.gz Velvet
-                  cd ../../..
-                fi;
+                $BB echo -e "${CYAN}  => Package installed ${NC}"
+                $BB echo -e "${CYAN}  => Backup Assistant Package ${NC}"
+                cp -f /cache/VelvetPrebuilt.tar.xz /sdcard/addon/VelvetPrebuilt.tar.xz
               else
-                busybox echo -e "${LRED}  => Package not installed ${NC}" && break
+                $BB echo -e "${LRED}  => Package not installed ${NC}" && break
               fi;
+              rm -rf /cache/VelvetPrebuilt.tar.xz 2>/dev/null;
+              $BB echo -e "${CYAN}  => Installation Finished ${NC}"
             else
-              busybox tar -xf /cache/prebuilt_Velvet.tar.xz -C /system/priv-app 2>/dev/null;
+              $BB tar -xf /cache/VelvetPrebuilt.tar.xz -C /system/priv-app 2>/dev/null;
               chmod 0755 /system/priv-app/Velvet 2>/dev/null;
               chmod 0644 /system/priv-app/Velvet/Velvet.apk 2>/dev/null;
               chcon -h u:object_r:system_file:s0 "/system/priv-app/Velvet/Velvet.apk" 2>/dev/null;
-              busybox echo -e "${CYAN}  => Check Installed Package ${NC}"
+              $BB echo -e "${CYAN}  => Check Installed Package ${NC}"
               if [ -f /system/priv-app/Velvet/Velvet.apk ]; then
-                busybox echo -e "${CYAN}  => Package installed ${NC}"
-                if [ -f /system/priv-app/Velvet/Velvet.apk ]; then
-                  busybox echo -e "${CYAN}  => Backup Assistant Package ${NC}"
-                  cd /system/priv-app
-                  busybox tar -czf /sdcard/addon/prebuilt_Velvet.tar.gz Velvet
-                  cd ../..
-                fi;
+                $BB echo -e "${CYAN}  => Package installed ${NC}"
+                $BB echo -e "${CYAN}  => Backup Assistant Package ${NC}"
+                cp -f /cache/VelvetPrebuilt.tar.xz /sdcard/addon/VelvetPrebuilt.tar.xz
               else
-                busybox echo -e "${LRED}  => Package not installed ${NC}" && break
+                $BB echo -e "${LRED}  => Package not installed ${NC}" && break
               fi;
             fi;
-            rm -rf /cache/prebuilt_Velvet.tar.xz 2>/dev/null;
-            busybox echo -e "${CYAN}  => Installation Finished ${NC}"
-            break
+            rm -rf /cache/VelvetPrebuilt.tar.xz 2>/dev/null;
+            $BB echo -e "${CYAN}  => Installation Finished ${NC}"
             ;;
         "Wellbeing")
-            busybox echo -e "${CYAN}  => Removing Pre-installed Wellbeing Package ${NC}"
-            rm -rf /system/priv-app/Wellbeing 2>/dev/null;
-            rm -rf /system/priv-app/WellbeingPrebuilt 2>/dev/null;
-            rm -rf /system/product/priv-app/Wellbeing 2>/dev/null;
-            rm -rf /system/product/priv-app/WellbeingPrebuilt 2>/dev/null;
-            busybox echo -e "${CYAN}  => Downloading Prebuilt Wellbeing Package ${NC}"
-            busybox wget https://github.com/BiTGApps/BiTGApps/raw/master/addon/prebuilt/prebuilt_Wellbeing.tar.xz -O /cache/prebuilt_Wellbeing.tar.xz
-            busybox echo -e "${CYAN}  => Installing Wellbeing Package ${NC}"
+            $BB echo -e "${CYAN}  => Downloading Prebuilt Wellbeing Package ${NC}"
+            curl https://bitgapps.cf/addon/WellbeingPrebuilt.tar.xz > /cache/WellbeingPrebuilt.tar.xz
+            $BB echo -e "${CYAN}  => Installing Wellbeing Package ${NC}"
             if [ "$android_sdk" == "29" ]; then
-              busybox tar -xf /cache/prebuilt_Wellbeing.tar.xz -C /system/product/priv-app 2>/dev/null;
+              $BB tar -xf /cache/WellbeingPrebuilt.tar.xz -C /system/product/priv-app 2>/dev/null;
               chmod 0755 /system/product/priv-app/WellbeingPrebuilt 2>/dev/null;
               chmod 0644 /system/product/priv-app/WellbeingPrebuilt/WellbeingPrebuilt.apk 2>/dev/null;
               chcon -h u:object_r:system_file:s0 "/system/product/priv-app/WellbeingPrebuilt/WellbeingPrebuilt.apk" 2>/dev/null;
-              busybox echo -e "${CYAN}  => Check Installed Package ${NC}"
+              $BB echo -e "${CYAN}  => Check Installed Package ${NC}"
               if [ -f /system/product/priv-app/WellbeingPrebuilt/WellbeingPrebuilt.apk ]; then
-                busybox echo -e "${CYAN}  => Package installed ${NC}"
-                if [ -f /system/product/priv-app/WellbeingPrebuilt/WellbeingPrebuilt.apk ]; then
-                  busybox echo -e "${CYAN}  => Backup WellbeingPrebuilt Package ${NC}"
-                  cd /system/product/priv-app
-                  busybox tar -czf /sdcard/addon/prebuilt_Wellbeing.tar.gz WellbeingPrebuilt
-                  cd ../../..
-                fi;
+                $BB echo -e "${CYAN}  => Package installed ${NC}"
+                $BB echo -e "${CYAN}  => Backup WellbeingPrebuilt Package ${NC}"
+                cp -f /cache/WellbeingPrebuilt.tar.xz /sdcard/addon/WellbeingPrebuilt.tar.xz
               else
-                busybox echo -e "${LRED}  => Package not installed ${NC}" && break
+                $BB echo -e "${LRED}  => Package not installed ${NC}" && break
               fi;
+              rm -rf /cache/WellbeingPrebuilt.tar.xz 2>/dev/null;
+              $BB echo -e "${CYAN}  => Installation Finished ${NC}"
             elif [ "$android_sdk" == "28" ]; then
-              busybox tar -xf /cache/prebuilt_Wellbeing.tar.xz -C /system/priv-app 2>/dev/null;
+              $BB tar -xf /cache/WellbeingPrebuilt.tar.xz -C /system/priv-app 2>/dev/null;
               chmod 0755 /system/priv-app/WellbeingPrebuilt 2>/dev/null;
               chmod 0644 /system/priv-app/WellbeingPrebuilt/WellbeingPrebuilt.apk 2>/dev/null;
               chcon -h u:object_r:system_file:s0 "/system/priv-app/WellbeingPrebuilt/WellbeingPrebuilt.apk" 2>/dev/null;
-              busybox echo -e "${CYAN}  => Check Installed Package ${NC}"
+              $BB echo -e "${CYAN}  => Check Installed Package ${NC}"
               if [ -f /system/priv-app/WellbeingPrebuilt/WellbeingPrebuilt.apk ]; then
-                busybox echo -e "${CYAN}  => Package installed ${NC}"
-                if [ -f /system/priv-app/WellbeingPrebuilt/WellbeingPrebuilt.apk ]; then
-                  busybox echo -e "${CYAN}  => Backup WellbeingPrebuilt Package ${NC}"
-                  cd /system/priv-app
-                  busybox tar -czf /sdcard/addon/prebuilt_Wellbeing.tar.gz WellbeingPrebuilt
-                  cd ../..
-                fi;
+                $BB echo -e "${CYAN}  => Package installed ${NC}"
+                $BB echo -e "${CYAN}  => Backup WellbeingPrebuilt Package ${NC}"
+                cp -f /cache/WellbeingPrebuilt.tar.xz /sdcard/addon/WellbeingPrebuilt.tar.xz
               else
-                busybox echo -e "${LRED}  => Package not installed ${NC}" && break
+                $BB echo -e "${LRED}  => Package not installed ${NC}" && break
               fi;
-              rm -rf /cache/prebuilt_Wellbeing.tar.xz 2>/dev/null;
-              busybox echo -e "${CYAN}  => Installation Finished ${NC}"
-              break
+              rm -rf /cache/WellbeingPrebuilt.tar.xz 2>/dev/null;
+              $BB echo -e "${CYAN}  => Installation Finished ${NC}"
             else
-              busybox echo -e "${LRED}  => Unsupported Android SDK Version $android_sdk ${NC}"
-              rm -rf /cache/prebuilt_Wellbeing.tar.xz 2>/dev/null;
-              busybox echo -e "${LRED}  => Package not installed ${NC}" && break
+              $BB echo -e "${LRED}  => Unsupported Android SDK Version $android_sdk ${NC}"
+              rm -rf /cache/WellbeingPrebuilt.tar.xz 2>/dev/null;
+              $BB echo -e "${LRED}  => Package not installed ${NC}" && break
             fi;
             ;;
         "Reboot")
-            busybox echo -e "${LRED}  => Rebooting in 2 Secs ${NC}"
+            $BB echo -e "${LRED}  => Rebooting in 2 Secs ${NC}"
             sleep 2
             reboot
             ;;
